match versions as strings
Aether (and Maven?) considers some versions equal even if they do
no have identical string representation. For example, 1.2.3.test
and 1.2.3-test are considered equal.

This was confusing the target platform editor and resulted in
multiple sematically equal and literally different versions
presented in the UI.

It is probably more correct to fix this in the editors, but much
easier to force target platform to do string/literal version
comparison.

Signed-off-by: Igor Fedorenko <igor@ifedorenko.com>